Refactored the user verification override
